move contexts + jsdoc
