+++ sales-coach-avatar/src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './env';
 export * from "./errorHandler";
 export * from "./logger";
 export * from "./types";
